+++ lib/detectable_text_field.dart
@@ -1,5 +1,10 @@
 library detectable_text_field;
 
-export 'package:detectable_text_field/detectable_text_field.dart';
+export 'package:detectable_text_field/widgets/detectable_text_field.dart';
+export 'package:detectable_text_field/widgets/detectable_text_editing_controller.dart';
+export 'package:detectable_text_field/widgets/detectable_text.dart';
+export 'package:detectable_text_field/hooks/use_detectable_text_editing_controller.dart';
+export 'package:detectable_text_field/detector/sample_regular_expressions.dart';
+export 'package:detectable_text_field/detector/text_pattern_detector.dart';
 
-export 'widgets/detectable_text.dart';
+export 'package:detectable_text_field/detectable_text_field.dart';