+++ lib/detector/sample_regular_expressions.dart
@@ -33,9 +33,9 @@ const detectionContentLetters = _symbols +
     _persianNumber +
     _thaiLetters;
 
-const urlRegexContent = "((http|https)://)(www.)?" +
-    "[a-zA-Z0-9@:%._\\+~#?&//=]" +
-    "{2,256}\\.[a-z]" +
+const urlRegexContent = "((http|https)://)?(www.)?" +
+    "[-a-zA-Z0-9@:%._\\+~#?&//=]" +
+    "{1,256}\\.[a-z]" +
     "{2,6}\\b([-a-zA-Z0-9@:%" +
     "._\\+~#?&//=]*)";