+++ CHANGELOG.md
@@ -1,27 +1,69 @@
-## [2.0.3] 
-- added read more function to DetectableText
-- added method to detect all the text
-- fixed problems example not being compiled
+## [3.0.2]
 
-## [2.0.2] 
-- fixed errors occurred on flutter 2.2.x
+- Fix example
 
-## [2.0.1] 
-- fixed error occurred in onDetectionFinished and onDetectionTyped.
+## [3.0.1]
 
-## [2.0.0] 
-- supported null safety
+- Fix analysis problems
 
-## [1.0.1] 
-- fixed the bug when auto correct suggestion appeared.
+## [3.0.0]
 
-## [1.0.0] 
-- fixed readme to show the update of v0.3.0
+- **Breaking**: Removed `onDetectionTyped` and `onDetectionFinished` from `DetectableTextField`.
+  Use `DetectableTextEditingController` instead. It is now possible to listen `typingDetection`
+  from `DetectableTextEditingController`.
+- DetectableTextField is no longer required. You can use `DetectableTextEditingController`
+  with `TextField` or `TextFormField`. DetectableTextField is just a extended TextField.
 
-## [0.3.0] - added sample regular expression of url detection
+## [2.1.0]
 
-## [0.2.0] - add `onDetectionTyped`
+- Supported flutter 3.10.0
+- Fixed the gestures of DetectableText
+- Fixed link detection
 
-## [0.1.0] - initial release
+## [2.0.4]
 
-## [0.1.0+1] - Fixed the style of README
+- Added contributors list on readme.
+- Fixed analysis problems
+
+## [2.0.3]
+
+- Added read more function to DetectableText
+- Added method to detect all the text
+- Fixed problems example not being compiled
+
+## [2.0.2]
+
+- Fixed errors occurred on flutter 2.2.x
+
+## [2.0.1]
+
+- Fixed error occurred in onDetectionFinished and onDetectionTyped.
+
+## [2.0.0]
+
+- Supported null safety
+
+## [1.0.1]
+
+- Fixed the bug when auto correct suggestion appeared.
+
+## [1.0.0]
+
+- Fixed readme to show the update of v0.3.0
+
+## [0.3.0]
+
+- Added sample regular expression of url detection
+
+## [0.2.0]
+
+- Added `onDetectionTyped`
+
+## [0.1.0+1]
+
+- Fixed the style of README
+
+## [0.1.0]
+
+- Initial release
+